fix: asset_summaries validation workaround, tests
